==> 1.Prepare_data/2. Prepare_user_job_data.ipynb <==
For each dataset, create job file, which contains
* JobID
* JobCity
* JobState
* JobCountry

For each dataset, create user file, which contains
* UserID
* UserCity
* UserState
* UserCountry
* UserDegree
* UserMajor

import random
import pickle
import argparse
import numpy as np
import pandas as pd
import tensorflow as tf

path =  "../data/"
dataset = "cb12/"

raw_path = path + dataset + "raw/" 
interim_path = path + dataset + "interim/"
processed_path = path + dataset + "processed/"

# Step 1: Load user data

df_user = pd.read_csv(raw_path + 'users.tsv', header=0, sep='\t')
print(df_user.shape)
df_user.dropna(axis=0, how='any', thresh=None, subset=['UserID', 'City', 'State', 'Country', 'DegreeType', 'Major', 'WorkHistoryCount', 'TotalYearsExperience'], inplace=True)
print(df_user.shape)

# Step 2: Load job data

### 30min

print('Loading train data from file: {}'.format(processed_path + 'train_14d_30_consider_user.csv'))
train_df = pd.read_csv(processed_path + 'train_14d_30_consider_user.csv', header=0, sep='\t')
train_df['All'] = train_df["Title"].map(str) + train_df["Description"].map(str) + train_df["Requirements"].map(str)
print('Train data shape: ', train_df.shape)

print('Loading test data from file: {}'.format(processed_path + 'test_14d_30_consider_user.csv'))
test_df = pd.read_csv(processed_path + 'test_14d_30_consider_user.csv', header=0, sep='\t')
test_df['All'] = test_df["Title"].map(str) + test_df["Description"].map(str) + test_df["Requirements"].map(str)
print('Test data shape: ', test_df.shape)

job_df = train_df.drop_duplicates(subset='JobID')
print(job_df.shape)

job_df.rename(columns={"City": "JobCity", "State": "JobState", "Country": "JobCountry"}, inplace = True)

job_df.to_csv(processed_path + 'jobs_14d_30_consider_user.csv', sep='\t',index=False)

all_df = pd.concat([train_df, test_df], axis = 0)
print(all_df.shape)

df_user_valid = df_user[df_user.UserID.isin(all_df.UserID)]
print(df_user_valid.shape)
df_user_valid.rename(columns={"City": "UserCity", "State": "UserState", "Country": "UserCountry", "DegreeType": "UserDegree", "Major": "UserMajor"}, inplace = True)

df_user_valid.to_csv(processed_path + 'users_14d_30_consider_user.csv', sep='\t',index=False)
