==> 1.Prepare_data/4.Tokenize.ipynb <==
Tokenize content and save

import os
import random
import pickle
import argparse
import matplotlib
import numpy as np
import pandas as pd
import tensorflow as tf

from tqdm import tqdm

path =  "../data/"
dataset = "cb12/"
raw_path = path + dataset + "raw/" 
interim_path = path + dataset + "interim/"
processed_path = path + dataset + "processed/"

# Step 1: Load job data

print('Loading job from file: {}'.format(processed_path + 'jobs_14d_30_consider_user_encoded.csv'))
job_df_30 = pd.read_csv(processed_path + 'jobs_14d_30_consider_user_encoded.csv', header=0, sep='\t')
print('Job data shape: ', job_df_30.shape)
print('Unique JobCity: ', len(job_df_30.JobCity.unique()))
print('Unique JobState: ', len(job_df_30.JobState.unique()))
print('Unique JobCountry: ', len(job_df_30.JobCountry.unique()))

# Step 2: Tokenize

print('Tokenizing texts and Converting tokens to int numbers...')

import nltk
from nltk import FreqDist
from nltk.corpus import stopwords
from string import digits, punctuation
from nltk.tokenize import word_tokenize
from gensim.models import KeyedVectors


def tokenize_text(text, clean_str_fn):
    text = clean_str_fn(text)
    tokenized_text = []
    if type(text) != float:
        text = text.strip(' ')
        remove_punctuations = str.maketrans('', '', punctuation)
        text = text.translate(remove_punctuations)
        remove_digits = str.maketrans('', '', digits)
        text = text.translate(remove_digits)
        tokens = text.split(' ')
        for token in tokens:
            if token and token not in stopwords.words('english'):
                tokenized_text.append(token)

    return tokenized_text


def tokenize_texts(texts, tokenization_fn=None, clean_str_fn=lambda x: x):
    if tokenization_fn == None:
        tokenized_texts = [tokenize_text(text, clean_str_fn) for text in texts]
    else:
        tokenized_texts = [tokenization_fn(text) for text in texts]

    return tokenized_texts


def get_words_freq(tokenized_texts):
    words_freq = FreqDist([word for text in tokenized_texts for word in text])
    return words_freq 

### Title

content_type = 'Title'
print('Tokenizing texts...')
tokenized_texts_title = tokenize_texts(job_df_30[content_type].values)
print('Computing word frequencies...')
# A dictionary 
words_freq_title = get_words_freq(tokenized_texts_title)
print('Number of vocabulary in {} (raw): {}'.format(content_type, len(words_freq_title)))

job_df_30['Title_tokenized'] = tokenized_texts_title

### Description

content_type = 'Description'
print('Tokenizing texts...')
tokenized_texts_description = tokenize_texts(job_df_30[content_type].values)
print('Computing word frequencies...')
# A dictionary 
words_freq_description = get_words_freq(tokenized_texts_description)
print('Number of vocabulary in {} (raw): {}'.format(content_type, len(words_freq_description)))

job_df_30['Description_tokenized'] = tokenized_texts_description

### Requirements

content_type = 'Requirements'
print('Tokenizing texts...')
tokenized_texts_requirements = tokenize_texts(job_df_30[content_type].values)
print('Computing word frequencies...')
# A dictionary 
words_freq_requirements = get_words_freq(tokenized_texts_requirements)
print('Number of vocabulary in {} (raw): {}'.format(content_type, len(words_freq_requirements)))

job_df_30['Requirements_tokenized'] = tokenized_texts_requirements

### All

content_type = 'All'
print('Tokenizing texts...')
tokenized_texts_all = tokenize_texts(job_df_30[content_type].values)
print('Computing word frequencies...')
# A dictionary 
words_freq_all = get_words_freq(tokenized_texts_all)
print('Number of vocabulary in {} (raw): {}'.format(content_type, len(tokenized_texts_all)))

job_df_30['All_tokenized'] = tokenized_texts_all

# Step 3: Save

job_df_30.to_csv(processed_path + 'jobs_14d_30_consider_user_encoded_tokenized.csv', sep='\t', index=False)
